create playlists
add facility to add item to playlist on the frontend. his includes creating a new playlist
or adding an item to an existing playlist. items can be added from the deck or the library

implemented creating a playlist, adding item to playlist, getting all playlists for a given user
and getting a specific playlist on the server side